+++ gtest/file_io.cc
@@ -11,10 +11,11 @@ class file_io : public ::testing::Test
  protected:
   void SetUp() override
   {
-    cuke::registry().push_step(
-        cuke::internal::step([](const cuke::value_array&) {}, "a step"));
     cuke::registry().push_step(cuke::internal::step(
-        [](const cuke::value_array&) {}, "a step with {int} and {string}"));
+        [](const cuke::value_array&, const auto&, const auto&) {}, "a step"));
+    cuke::registry().push_step(cuke::internal::step(
+        [](const cuke::value_array&, const auto&, const auto&) {},
+        "a step with {int} and {string}"));
   }
   void TearDown() override
   {