+++ gtest/run_scenario_hooks.cc
@@ -22,7 +22,8 @@ class run_scenario_hooks : public ::testing::Test
     cuke::registry().push_hook_before_step(hook([]() { ++before_step_calls; }));
     cuke::registry().push_hook_after_step(hook([]() { ++after_step_calls; }));
 
-    cuke::registry().push_step(step([](const cuke::value_array&) {}, "a step"));
+    cuke::registry().push_step(step(
+        [](const cuke::value_array&, const auto&, const auto&) {}, "a step"));
   }
 
   static std::size_t before_calls;
@@ -83,11 +84,13 @@ class run_scenario_hook_skip : public ::testing::Test
 
     cuke::registry().clear();
 
-    cuke::registry().push_step(
-        step([](const cuke::value_array&) { ++calls; }, "a step"));
+    cuke::registry().push_step(step(
+        [](const cuke::value_array&, const auto&, const auto&) { ++calls; },
+        "a step"));
 
-    cuke::registry().push_step(
-        step([](const cuke::value_array&) { ++calls; }, "a step with {int}"));
+    cuke::registry().push_step(step(
+        [](const cuke::value_array&, const auto&, const auto&) { ++calls; },
+        "a step with {int}"));
 
     cuke::registry().push_hook_before(
         hook([]() { cuke::skip_scenario(); }, "@skip"));
@@ -235,7 +238,7 @@ TEST_F(run_scenario_hook_skip, run_taggeed_example)
     Examples: 
     | value | 
     | 1     |
-  )*";
+ )*";
 
   cuke::parser p;
   p.parse_script(script);