+++ gtest/test_results.cc
@@ -12,10 +12,11 @@ class test_results_1 : public ::testing::Test
   {
     cuke::results::test_results().clear();
     cuke::registry().clear();
-    cuke::registry().push_step(
-        cuke::internal::step([](const cuke::value_array&) {}, "a step"));
     cuke::registry().push_step(cuke::internal::step(
-        [](const cuke::value_array&) { cuke::is_true(false); }, "this fails"));
+        [](const cuke::value_array&, const auto&, const auto&) {}, "a step"));
+    cuke::registry().push_step(cuke::internal::step(
+        [](const cuke::value_array&, const auto&, const auto&)
+        { cuke::is_true(false); }, "this fails"));
   }
 };